Add initial split of executor and listener
